refactor(chat.jsx): create ChatInput.jsx comp and refactor Chat.jsx comp - Add ChatInput.jsx comp for handling messaging functionality - Remove duplicate code from Chat.jsx comp - Render only chat inside Chat.jsx comp
